jobPostingRepositoryImpl 합침
